pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "murmur-voice"
version = "0.1.0"
description = "Local-first agentic voice pipeline: STT → LLM → TTS with every major provider"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "Murmur Contributors" }]
keywords = [
  "voice-assistant", "speech-to-text", "text-to-speech", "whisper",
  "openrouter", "llm", "local-ai", "self-hosted", "agentic-ai",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
  "click>=8.1",
  "rich>=13.0",
  "tomli>=2.0; python_version < '3.11'",
  "httpx>=0.27",
  "openai>=1.40",          # OpenRouter uses OpenAI-compatible API
  "sounddevice>=0.4.6",
  "numpy>=1.26",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "aiofiles>=23.0",
]

[project.optional-dependencies]
# STT providers
faster-whisper = ["faster-whisper>=1.0"]
deepgram = ["deepgram-sdk>=3.0"]
assemblyai = ["assemblyai>=0.25"]
groq-stt = ["groq>=0.9"]
openai-stt = []  # uses openai already in core

# TTS providers
kokoro = ["kokoro>=0.9; sys_platform != 'win32'", "soundfile>=0.12"]
piper = ["piper-tts>=1.2"]
elevenlabs = ["elevenlabs>=1.0"]
cartesia = ["cartesia>=1.0"]
edge-tts = ["edge-tts>=6.1"]
openai-tts = []  # uses openai already in core

# LLM extras
anthropic = ["anthropic>=0.30"]

# VAD
vad = ["silero-vad>=5.0"]
webrtcvad = ["webrtcvad-wheels>=2.0"]

# UI
ui = ["fastapi>=0.111", "uvicorn[standard]>=0.30", "websockets>=12.0"]

# Dev
dev = [
  "pytest>=8.0", "pytest-asyncio>=0.23", "pytest-cov>=5.0",
  "ruff>=0.5", "black>=24.0", "mypy>=1.10",
  "pre-commit>=3.7", "httpx>=0.27",
]

# All local (no API keys needed)
local = [
  "murmur-voice[faster-whisper,kokoro,vad]",
]

# All cloud providers
cloud = [
  "murmur-voice[deepgram,assemblyai,groq-stt,elevenlabs,cartesia,edge-tts,anthropic]",
]

# Everything
all = [
  "murmur-voice[local,cloud,ui]",
]

[project.scripts]
murmur = "murmur.cli:cli"

[project.urls]
Homepage = "https://github.com/murmur-voice/murmur"
Documentation = "https://murmur.dev/docs"
Repository = "https://github.com/murmur-voice/murmur"
Issues = "https://github.com/murmur-voice/murmur/issues"
Discord = "https://discord.gg/murmur"

[tool.hatch.build.targets.wheel]
packages = ["murmur"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["murmur"]
omit = ["tests/*"]
